requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
langchain-text-splitters
PyMuPDF
pypdf
chromadb
sentence-transformers
python-dotenv

